Inline data loading into the login preferences component
